+++ docs/releases.md
@@ -4,11 +4,14 @@
 
 ### New Features
 
+* Support to plot ROMS output data at lat/lon locations ([#277](https://github.com/CWorthy-ocean/roms-tools/pull/277))
+
 ### Breaking Changes
 
 ### Internal Changes
 
 * Unfreeze arguments in all dataclasses ([#276](https://github.com/CWorthy-ocean/roms-tools/pull/276))
+* Integration with xesmf ([#277](https://github.com/CWorthy-ocean/roms-tools/pull/277))
 
 ### Documentation